Replace Flask-Login with session/decorators
Just in case, Flask-Login has been replaced with the use of Flask's
session variable and decorators, as well as changing some of the
database calls.

This is to avoid any possibility of getting docked for using ORM
features, since Flask-Login requires a User class of some sort,
even if that User class doesn't specifically use ORM features of
some database module; this is to be completely safe.